pyproject.toml
[project]
name = "stock-project"
version = "0.1.0"
description = "주식 리포트 자동화 시스템"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "markdown>=3.6",
    "notion-client>=2.2.1",
    "pytrends>=4.9.2",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "feedparser>=6.0.12",
    "google-genai>=1.65.0",
    "aiohttp>=3.13.3",
    "tenacity>=9.1.4",
    "playwright>=1.58.0",
    "fastapi>=0.134.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.3.0",
]

[tool.ruff]
# 파이썬 표준에 맞춘 줄바꿈 길이
line-length = 88
target-version = "py313"

[tool.ruff.lint]
# E, W: PEP8 에러/경고, F: Pyflakes, I: isort, D: pydocstyle
select = ["E", "W", "F", "I", "D"]

# pydocstyle 설정을 Google 스타일로 명시
[tool.ruff.lint.pydocstyle]
convention = "google"
